// src/app/models/booking.ts
// ng generate interface models/booking
export interface IBooking {
  id?: number; // Optional, as it's generated by JSON Server on POST
  name: string;
  city: string;
  phone: string;
  email: string;
  age: number; // Changed to number for premium calculation
  planId: number;
  planName: string;
  validity: number; // e.g., 30, 60, 365
  paymentMode: string;
  cardNumber: string;
  premiumAmt: number; // The calculated final amount
  paymentFreq: 'yearly' | 'half-yearly' | 'monthly'; // Example values
}